Refactor: 비즈니스 정책 변경에 따른 패키지 구조 변경
### 변경 전

ex) web/customer/me...

### 변경 후

ex) web/domain/customer/...
