+++ src/main/java/com/prgms/allen/dining/web/domain/customer/me/CustomerMyReservationApi.java
@@ -1,9 +1,10 @@
-package com.prgms.allen.dining.web.customer.me;
+package com.prgms.allen.dining.web.domain.customer.me;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/customer/api/me")
-public class CustomerMyPageApi {
+@RequestMapping("/customer/api/me/reservations")
+public class CustomerMyReservationApi {
+
 }
