+++ src/main/java/com/prgms/allen/dining/web/domain/customer/reservation/CustomerReservationApi.java
@@ -1,0 +1,9 @@
+package com.prgms.allen.dining.web.domain.customer.reservation;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/customer/api/reservations")
+public class CustomerReservationApi {
+}
