+++ src/main/java/com/prgms/allen/dining/web/domain/customer/restaurant/CustomerRestaurantApi.java
@@ -1,10 +1,9 @@
-package com.prgms.allen.dining.web.customer.me;
+package com.prgms.allen.dining.web.domain.customer.restaurant;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/customer/api/me/reservations")
-public class CustomerMyReservationApi {
-
+@RequestMapping("/customer/api/restaurants")
+public class CustomerRestaurantApi {
 }
