+++ src/main/java/com/prgms/allen/dining/web/domain/owner/reservation/OwnerReservationApi.java
@@ -1,4 +1,4 @@
-package com.prgms.allen.dining.web.owner.reservation;
+package com.prgms.allen.dining.web.domain.owner.reservation;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
