+++ src/main/java/com/prgms/allen/dining/web/domain/owner/restaurant/OwnerRestaurantApi.java
@@ -1,9 +1,9 @@
-package com.prgms.allen.dining.web.customer.reservation;
+package com.prgms.allen.dining.web.domain.owner.restaurant;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/customer/api/reservations")
-public class CustomerReservationApi {
+@RequestMapping("/owner/api/restaurants")
+public class OwnerRestaurantApi {
 }
