--- src/main/java/com/prgms/allen/dining/web/owner/me/OwnerMyPageApi.java
@@ -1,9 +1,0 @@
-package com.prgms.allen.dining.web.owner.me;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/owner/api/me")
-public class OwnerMyPageApi {
-}
